Bug fix. generate_fixture command fails when you have a GeoDjango app or some other specific database backend magic that sqlite3 is not able to reproduce. Therefore syncdb command fails at generating the tables in the database.
I have added a --default option that will use default project settings
and generate test databases, that will use for populating the fixtures.

TODO: Maybe it's a good idea that the user can provide a test_settings
file as a parameter. So he can use specific settings for generating
the databases.